created content/subject for component and spec folder. Moved relevant files and edited component path where rendered
