+++ app/components/content/subject/subject_page_content_component.html.erb
