+++ app/views/content/life-as-a-teacher/explore-subjects/biology/_article.html.erb
@@ -3,18 +3,9 @@
     <h2 class="heading--box-blue">Why teach biology</h2>
     <div class="inset">
       <p>Biology is the study of living things, the intricate systems that sustain them, and the delicate balance of our ecosystems. As a biology teacher, you'll have the opportunity to inspire the next generation to explore the wonders of life itself.</p>
-      <%
-=begin%>
- <p>In addition to the personal rewards of teaching biology, there are practical benefits too, including:</p>
-      <ul>
-        <li>a <a href="/life-as-a-teacher/pay-and-benefits/teacher-pay">competitive salary starting at <%= value :salaries_starting_minshortened %></a> (or higher in London)</li>
-        <li>a <a href="/life-as-a-teacher/pay-and-benefits/teachers-pension-scheme">generous and secure pension</a></li>
-        <li>more holiday days than people in many other professions</li>
-        <li>opportunities for progression</li>
-      </ul> 
-<%
-=end%>
-        <%= render Content::SubjectPageContentComponent.new(subject_name: "biology") %>
+
+        <%= render(Content::Subject::SubjectPageContentComponent.new(subject_name: "biology")) %>
+
     </div>
   </section>
